latch on to activation hook for initializing capabilities
