+++ public/class-bigbluebutton-public.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * The public-facing functionality of the plugin.
+ * The views of the plugin.
  *
  * @link       https://blindsidenetworks.com
  * @since      3.0.0
@@ -11,7 +11,7 @@
  */
 
 /**
- * The public-facing functionality of the plugin.
+ * The views of the plugin.
  *
  * Defines the plugin name, version, and two examples hooks for how to
  * enqueue the public-facing stylesheet and JavaScript.
